requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
click==8.1.3
Django==4.1.3
django-ninja==0.19.1
h11==0.14.0
pydantic==1.10.2
sqlparse==0.4.3
typing-extensions==4.4.0
uvicorn==0.19.0
pandas==1.5.1
psycopg2-binary==2.8.4